docker-storage-setup: Grow rootfs and pvs early in the process
Grow root pvs and rootfs early in the process. That seems to be completely
orthogonal to setting up an lvm thin pool. In fact it should be part of
a separate serivce.

This seems to be the only thing which needs to run on every reboot. Otherwise
once lvm thin pool has been setup, there does not seem to be a need for this
service to run on every boot.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>