docker-storage_setup: Do not create thin pool if storage is already configured
This service creates a thin pool and writes parameters to docker in
/etc/sysconfig/docker-storage file in option DOCKER_STORAGE_OPTIONS.

It is possible that user has hand edited the /etc/sysconfig/docker-storage
to use overlay. Or to provide additional storage options to docker daemon
or to use a custom thin pool user created. Right now re-running the service
over reboot might override user's configuration.

So don't create thin pool and overwrite this file if DOCKER_STORAGE_OPTONS=
is not nil. 

User will have to cleanup this file manually and also cleanup any associated
configuraiton in docker and then re-run the script/service.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>