docker-storage-setup: Add a config option for docker root
Add a config option DOCKER_ROOT_VOLUME which specifies whether this script
should setup a separate volume for docker root directory or not. By default
it is turned off.

This can be helpful when using overlay driver and one does not want all
the images and containers on rootfs. 

In that case one can prepare a separate volume which will be mounted on
directory specified by DOCKER_ROOT_DIR and docker will be asked to use
DOCKER_ROOT_DIR as root dir (using option -g).
 
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>