Wait for previously configured thin pool to come up
If docker-storage-setup has created a thin pool then on next boot it is
possible that thin pool, volume group and underlying pvs are on a slow
storage and are not available when docker-storage-setup runs. Right now
docker-sotrage-setup will assume thin pool does not exist and might error
out or try to create new thin pool based on configuration.

This patch adds logic to wait for the thin pool to come up till a certain
timeout period. This wait happens only if we find that /etc/sysconfig/docker-storage contains a thin pool specification using dm.thinpooldev=.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>