Use pvcreate with option -f
By the time we run pvcreate, we have already checked the disk for any
partition table or any other signatures present. That should mean that
we own the disk now and it is save to use it. If some signatures are
found on disk after creating partition (wipefs /dev/sda1), then we
should be able to ignore it.

IOW, if user creates are partition say /dev/sda1 and does mkfs.xfs /dev/sda1,
and later removes partition /dev/sda1 and passes disk to dss, then 
xfs signatures created on /dev/sda1 will be ignored. We will care for
signatures only on whole disk that is /dev/sda.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>