Do not use pvcreate -f option
Do not use pvcreate -f option instead wipe signatures explicitly on the
partition device. Reason being that -f is not very well defined and 
tomorrow it can be overloaded to do more things. So it is safer to wipe
signatures explicitly and use pvcreate without -f.

At this point we own the disk. If there are any signatures on disk to
being with, we will not be here until and unless user asked to wipe
signatures using WIPE_SIGNATURES=true. We will not care for any signatures
found on disk after creating partition.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>