Do not try to manage thin pool not created by dss
In general /etc/sysconfig/docker-storage is managed by dss and user
should not edit it directly.

In the past there have been cases where this was not a strict
requirement. And looks like over upgrade we ran into a case
where a thin pool configured by user is not in same format
as dss expects it to be and various assumptions are broken
and errors happen.

Try to parse thin pool name and make sure it is in same
format as generated by dss and if it is not, exit early
and stop doing further processing.

This should allow user to continue with its setup.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>