Return error if command in subshell fails
If a command in sub-shell failed, return with error. Otherwise sometimes
it can continue if parent is capturing the error code of called funciton.

IOW, set -e makes sure that script will exit if command in sub-shell
failed but that will happen only if caller was not capturing error
code of called function.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>