Fix sfdisk failure on older sfdisk
We used to use following input file for sfdisk.

units:sectors
{dev}1: start=2048, size=4096, Id=8e

And this worked both with old and new sfdisk. But this required
hardcoding partition name and we wanted to move away from that
as it did not work for pmem and loop devices. So we started using
following format instead.

units:sectors
start=2048, size=4096, Id=8e

As per latest sfdisk man page this should work but it does not seem to
work with sfdisk shipped with older util-linux (util-linux-2.23.2-33),
and we get errors like following.

sfdisk: trailing junk after number

sfdisk: bad input

But following format seems to work with old sfdisk format. So switch
to using that one.

units:sectors
2048, 4096, 8e

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>