Do not resolve symlink device name if it is /dev/mapper/* device
Following PR requested resolving device symlinks.

https://github.com/projectatomic/container-storage-setup/pull/168

Primary problem seemd to be sylink renamed upon reattaching the device in
openstack. 

I want to provide an exception to /dev/mapper/* devices which are symlinks
but we don't want to resolve these as these links are persistent over
reboot and all tools work with these and lsblk even report same device
names (instead of /dev/dm* devices). It is recommended to work with
these links instead of devices in /dev/dm*.  So I am skipping resolving
symlinks for devices in /dev/mapper/ dir.

I need this to figure out if a multipath device is partition or not. I
rely on lsblk output and it reports /dev/mapper/ devices not the one
in /dev/. And that means my string check logic does not work. So do
not resolve symlink for /dev/mapper/* devices.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>