Make partition check work for dm multipath devices
Currently, is_block_dev_partiton() check fails if a multipath device 
is passed in. We look for parent of a device and if parent is present
we assume device must have been partitioned. 

While this is true for regular block devices, it is not true for stacked
blocked devices (dm, md, lvm etc). 

So refine the check further. Try to find parent device using lsblk and
if parent device name is substring of passed in block device, then most
likely passed in device is a partition.

For example, following is partial output of multipath partition device.

$lsblk -nslp /dev/mapper/mp1
/dev/mapper/mp1 253:3    0   6G  0 dm   
/dev/mapper/mp  253:2    0   6G  0 dm   
/dev/nvme0n1p1  259:5    0   6G  0 part 
/dev/nvme0n1    259:1    0   6G  0 disk 

Here parent device of /dev/mapper/mp1 is /dev/mapper/mp and parent is
a substring so it is fair to conclude that /dev/mapper/mp1 is a
partition device.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>